apps/backend/pyproject.toml
[tool.poetry]
name = "covsonar-backend"
version = "0.5.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "covsonar_backend"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
django = "^4.2.7"
numpy = "^1.26.1"
pandas = "^2.1.2"
django-redis = "^5.4.0"
django-cors-headers = "^4.3.0"
djangorestframework = "^3.14.0"
django-filter = "^23.3"
django-apscheduler = "^0.6.2"
psycopg2-binary = "^2.9.9"
biopython = "^1.81"
django-debug-toolbar = "^4.2.0"
whitenoise = "^6.6.0"
django-environ = "^0.11.2"
celery = {extras = ["redis"], version = "^5.3.6"}
dnspython = "^2.6.1"
parameterized = "^0.9.0"
black = "^24.3.0"
flower = "^2.0.1"
json-log-formatter = "^1.0"
gunicorn = "^22.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
